filter also could receive the index as the second parameter.
So that the FirstOp now reusing the filter operation.
The converge method no longer get the result, the result would be got in
evaluate method since converge method might be used to redirect the data
flow and the downstream is not a terminate pond.
canal.first method has been rename to top.
canal.first method now is a terminate operation which returns an Option
object might contain the first data.